refactor: 型定義をtypeに統一した。
