+++ app/src/components/Header.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import React from 'react';
 import styled from 'styled-components';
 
-import { SearchJson } from '@/interfaces/search-posts';
+import { SearchJson } from '@/types/search-posts';
 
 import { device } from '@/share/media-query';
 import { colorObj } from '@/share/variables';
@@ -11,7 +11,7 @@ import { colorObj } from '@/share/variables';
 import HeaderNav from '@/components/HeaderNav';
 import SearchArea from '@/components/Search/SearchArea';
 
-interface Props {
+type Props = {
   siteTitle: React.ReactNode;
   searchPosts: SearchJson[];
 }
