+++ app/src/types/page-slug.ts
@@ -1,4 +1,4 @@
 // getStaticPathsで取得したスラッグが格納されたオブジェクトの型
-export interface PageSlug {
+export type PageSlug = {
   id: string;
-}
+};
