+++ app/src/utils/generate-build-paginate-path.ts
@@ -1,4 +1,4 @@
-import { BuildTaxonomyPaginateList } from '@/interfaces/taxonomy';
+import { BuildTaxonomyPaginateList } from '@/types/taxonomy';
 import { config } from '@/config/app';
 
 export const generateBuildPaginatePath = (contents: BuildTaxonomyPaginateList[]) => {
